// src/app/Models/Data.model.ts
export class Detail {
  serialId: string = '';
  fullName: string = '';
  date: any;
  purpose: string = '';
  totalAmount: string = '';
  tAmount: string = '';
}

export class AddDetails {
  serialId: string = '';
  fullName: string = '';
  date: any;
  purpose: string = '';
  totalAmount: string = '';
  tAmount: string = '';
}

export class UpdateDetails {
  fullName: string = '';
  date: any;
  purpose: string = '';
  totalAmount: string = '';
  tAmount: string = '';
}